predict_microservice/requirements.txt
google-cloud-storage==1.42.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Werkzeug==0.16.1
waitress==2.0.0
Pillow==8.3.1
tensorflow==2.6.0
opencv-python==4.5.3.56
numpy
